Add export_custom.py for unified ONNX export of YOLO and AHOY models
This new script consolidates the export functionality for both YOLO and AHOY models into a single entry point. It allows users to specify detection weights and optional horizontal weights, streamlining the process of exporting models to ONNX format. The previous separate export scripts for YOLO and AHOY have been removed to reduce redundancy and improve maintainability.
